requirements-customizations.txt
git+https://github.com/peppelinux/pysaml2.git@pplnx-v7.4.2

#Bootstrap Italia + Unical template
git+https://github.com/italia/design-django-theme.git
git+https://github.com/UniversitaDellaCalabria/django-unical-agid-template.git
# SASS dynamic compiler
libsass
django_sass_processor

# ldap
git+https://github.com/peppelinux/django-ldapdb.git
git+https://github.com/peppelinux/pySSHA-slapd.git
git+https://github.com/peppelinux/django-ldap-academia-ou-manager.git
git+https://github.com/peppelinux/pyMultiLDAP.git
ldap3

# diagrams
# django-extensions
# pydotplus
# git+https://github.com/silentsokolov/django-admin-rangefilter.git

# custom maccounts in example project
pycountry

# oas3
djangorestframework
# Markdown support for the browsable API.
markdown     
django-filter
uritemplate
django-cors-headers # for rest api

# proxy
django-allauth>0.58,<0.59

cryptojwt>=1.8,<=1.9

django-cleanup>=8.0,<9.0

# password_reset
django-form-builder>=1.0.1 ,<1.1.0
